changes/2020/07/2020-07-28.html

<!DOCTYPE html>
<!--This is auto-generated file, don't edit-->
<html>
<head>
  <meta charset="UTF-8">
  <meta name="viewport" content="width=device-width"/>
  <title>Firefox Nightly changes 2020-07-28</title>
  <link href="../../../css/main.css" rel="stylesheet" type="text/css"/>
</head>
<body>
  <h2>Changes in 20200728094725</h2>
  <h3>Built from eba7e3ce93822075543bcc764cef7dbc8e9fc5f1</h3>
  <div id="bugList">
  <div class="logline even"><a href="index.html">..index..</a></div>
  <div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654383" title="NEW">Bug 1654383 N</a><span>[Core:Audio/Video: GMP]</span><span>Use encryption scheme enum from content_decryption_module.h rather than re-implementing as a GMP type</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655175" title="RESOLVED">Bug 1655175 R</a><span>[Core:Audio/Video: Playback]</span><span>Crash in [@ mozilla::dom::HTMLMediaElement::MediaControlKeyListener::InitMediaAgent]</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654195" title="RESOLVED">Bug 1654195 R</a><span>[Core:DOM: Animation]</span><span>Make text-combine-upright non-animatable</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655398" title="NEW">Bug 1655398 N</a><span>[Core:DOM: CSS Object Model]</span><span>`document.documentElement.clientTop` return an incorrect value</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1651357" title="ASSIGNED">Bug 1651357 A</a><span>[Core:DOM: Core & HTML]</span><span>Intermittent dom/base/test/browser_force_process_selector.js | final tab cannot be in its own process -</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655248" title="RESOLVED">Bug 1655248 R</a><span>[Core:DOM: Core & HTML]</span><span>Perma /html/cross-origin-opener-policy/reporting/access-reporting/property.https.html | cross-origin > w => w.frames - assert_not_equals: Report not received got disallowed value "timeout"</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653534" title="RESOLVED">Bug 1653534 R</a><span>[Core:DOM: Editor]</span><span>`WSRunScanner` and `TextFragmentData` should not refer style of scan start container for checking whether specific node is preformatted or not</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654533" title="RESOLVED">Bug 1654533 R</a><span>[Core:DOM: Security]</span><span>Intermittent TEST-UNEXPECTED-PASS | /content-security-policy/inheritance/iframe-all-local-schemes-inherit-self.sub.html | &lt;iframe>'s about:blank inherits policy. - expected TIMEOUT</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654088" title="RESOLVED">Bug 1654088 R</a><span>[Core:DOM: Window and Location]</span><span>window.open() truncates URL at first null byte</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654679" title="RESOLVED">Bug 1654679 R</a><span>[Core:Disability Access APIs]</span><span>Add selected text marker range to text selection changed events</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1648813" title="NEW">Bug 1648813 N</a><span>[Core:Disability Access APIs]</span><span>High frequency accessible/tests/browser/mac/browser_webarea.js | Test timed out -</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654169" title="RESOLVED">Bug 1654169 R</a><span>[Core:Graphics]</span><span>Sync surface of GeckoSurface always uses non single buffer surface</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1623792" title="NEW">Bug 1623792 N</a><span>[Core:Graphics: WebRender]</span><span>Optimize primitive dependency handling, taking advantage of spatial information from tile assignment.</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655472" title="RESOLVED">Bug 1655472 R</a><span>[Core:JavaScript Engine]</span><span>Calculate top level extent in jsparagus</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653567" title="RESOLVED">Bug 1653567 R</a><span>[Core:JavaScript Engine]</span><span>Redesign Private Fields implementation to reduce complexity</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654286" title="RESOLVED">Bug 1654286 R</a><span>[Core:JavaScript Engine]</span><span>Setting a pref in js/xpconnect/tests/unit/xpcshell.ini doesn't apply</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654495" title="RESOLVED">Bug 1654495 R</a><span>[Core:JavaScript Engine: JIT]</span><span>Support more than 256 CacheIR ops</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655203" title="ASSIGNED">Bug 1655203 A</a><span>[Core:JavaScript: GC]</span><span>Add tests for AutoStableStringChars and deduplication</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1651815" title="RESOLVED">Bug 1651815 R</a><span>[Core:JavaScript: Standard Library]</span><span>Stop using CallNonGenericMethod for Date methods</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1649718" title="RESOLVED">Bug 1649718 R</a><span>[Core:Javascript: WebAssembly]</span><span>Tweak compiler selection to cleanup + benefit fuzzing</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1651799" title="RESOLVED">Bug 1651799 R</a><span>[Core:Javascript: WebAssembly]</span><span>Simplify switches and flags for features that have shipped</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655430" title="RESOLVED">Bug 1655430 R</a><span>[Core:Layout: Form Controls]</span><span>De-duplicate MakeAnonymousElement from nsSearchControlFrame and nsNumberControlFrame</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=558594" title="RESOLVED">Bug 558594 R</a><span>[Core:Layout: Form Controls]</span><span>Layout for &lt;input type="search"></span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655237" title="RESOLVED">Bug 1655237 R</a><span>[Core:Panning and Zooming]</span><span>AsyncPanZoomController::HandleDragEvent wrongly clears the state for all mouse ups</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655474" title="NEW">Bug 1655474 N</a><span>[Core:Print Preview]</span><span>Crash in [@ nsPrintJob::UpdateSelectionAndShrinkPrintObject]</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654982" title="RESOLVED">Bug 1654982 R</a><span>[Core:Printing: Output]</span><span>Silent printing doesn't work on Mac - Firefox 78.x</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653667" title="RESOLVED">Bug 1653667 R</a><span>[Core:Web Audio]</span><span>SIGXCPU in [@ CallProcess]</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1651703" title="RESOLVED">Bug 1651703 R</a><span>[Core:Web Audio]</span><span>Update AudioParamDescriptor webidl min and max values</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1651601" title="RESOLVED">Bug 1651601 R</a><span>[Core:WebRTC]</span><span>Intermittent Assertion failed: (0), function nr_ice_media_stream_start_checks, file /builds/worker/checkouts/gecko/media/mtransport/third_party/nICEr/src/ice/ice_media_stream.c, line 414.</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1643157" title="RESOLVED">Bug 1643157 R</a><span>[Core:Widget: Cocoa]</span><span>Styling OPTION elements with background-color causes scrollbar to appear on MacOS</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654476" title="RESOLVED">Bug 1654476 R</a><span>[Core:Widget: Gtk]</span><span>[Linux/Gtk] Firefox fails to paste text/html mime if content="text/html; is missing</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654483" title="RESOLVED">Bug 1654483 R</a><span>[Core:XPCOM]</span><span>AssertionError: No line resembling a stack frame was found</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1651402" title="RESOLVED">Bug 1651402 R</a><span>[DevTools:Console]</span><span>Switch devtools to use Debugger.Object.isError instead of .class</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1644649" title="RESOLVED">Bug 1644649 R</a><span>[DevTools:General]</span><span>Intermittent devtools/server/tests/browser/browser_watcher-watchTargets-frames.js | Test timed out -</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654468" title="ASSIGNED">Bug 1654468 A</a><span>[DevTools:General]</span><span>Permafailing tier 2 devtools/client/application/test/browser/browser_application_panel_telemetry-start-worker.js | Test timed out -</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653907" title="RESOLVED">Bug 1653907 R</a><span>[DevTools:Inspector]</span><span>Remove outdated DevTools style actor trait getAppliedCreatesStyleCache</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1620264" title="RESOLVED">Bug 1620264 R</a><span>[DevTools:Inspector: Rules]</span><span>Intermittent devtools/client/inspector/rules/test/browser_rules_media-queries_reload.js | Uncaught exception - chrome://mochitests/content/browser/devtools/client/inspector/rules/test/head.js:805 - Error: Declaration color:red not found on rule at index 1</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654439" title="RESOLVED">Bug 1654439 R</a><span>[Firefox:Address Bar]</span><span>Subclass SearchOneOffs for urlbar</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654672" title="RESOLVED">Bug 1654672 R</a><span>[Firefox:Disability Access]</span><span>Site zoom scalar got removed but code is still setting it</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1635657" title="RESOLVED">Bug 1635657 R</a><span>[Firefox:Firefox Accounts]</span><span>Synthesize placeholder AET identifiers for existing accounts</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655390" title="RESOLVED">Bug 1655390 R</a><span>[Firefox:Firefox Accounts]</span><span>ensureProfile should handle missing profile cache</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654827" title="RESOLVED">Bug 1654827 R</a><span>[Firefox:PDF Viewer]</span><span>Update pdf.js to version 2.6.136</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653958" title="RESOLVED">Bug 1653958 R</a><span>[Firefox:Preferences]</span><span>Search result not correctly aligned for Zoom section when searching for numbers in about:preferences</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655044" title="RESOLVED">Bug 1655044 R</a><span>[Firefox:Protections UI]</span><span>[Protections Dashboard] Replace FPN Logo</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1377100" title="RESOLVED">Bug 1377100 R</a><span>[Firefox:Sync]</span><span>Passwords don't sync because sync never recovers from master password reset / failure to decrypt logins / corrupt/missing logins.json</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655340" title="RESOLVED">Bug 1655340 R</a><span>[Firefox Build System:General]</span><span>Cannot find autoconf 2.13</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655371" title="RESOLVED">Bug 1655371 R</a><span>[Firefox Build System:General]</span><span>Intermittent check> z:\build\build\src\ipc\ipdl\test\ipdl\ok\Punion_Comparable.ipdl:9: error: Unknown sync IPC message Punion_Comparable::Msg</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654465" title="RESOLVED">Bug 1654465 R</a><span>[Firefox Build System:General]</span><span>7.57 - 9.16% build times (linux64-shippable, osx-shippable) regression on pushrange 034864174a - bb79da633a (Mon July 20 2020)</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1644409" title="UNCONFIRMED">Bug 1644409 U</a><span>[Firefox Build System:General]</span><span>rustc link time optimization option is "fat"</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655583" title="RESOLVED">Bug 1655583 R</a><span>[Firefox Build System:General]</span><span>make[2]: warning: jobserver unavailable: using -j1.  Add '+' to parent make rule.</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1652503" title="RESOLVED">Bug 1652503 R</a><span>[Firefox Build System:Lint and Formatting]</span><span>Perma 0:29.34 tools/lint/test/test_black.py::test_lint_black[black] TEST-UNEXPECTED-FAIL</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654882" title="RESOLVED">Bug 1654882 R</a><span>[Firefox Build System:Task Configuration]</span><span>build makecab in automation, rather than getting it from tooltool</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655499" title="RESOLVED">Bug 1655499 R</a><span>[Firefox Build System:Task Configuration]</span><span>./mach try fuzzy with `rebuild` argument fails the decision task</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654988" title="RESOLVED">Bug 1654988 R</a><span>[Firefox Build System:Task Configuration]</span><span>Intermittent taskgraph.util.bugbug.BugbugTimeoutException: Timed out waiting for result from 'https://bugbug.herokuapp.com/push/integration/autoland/f6dc8b4c1814928ee3b8bfe1ea33a14ce9ae2328/schedules'</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654621" title="RESOLVED">Bug 1654621 R</a><span>[GeckoView:General]</span><span>[1530022 subtask 1] Refactor nsIExternalHelperAppService interface</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653171" title="ASSIGNED">Bug 1653171 A</a><span>[Release Engineering:Release Automation: L10N]</span><span>Timeouts cloning l10n repositories in windows repacks</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654167" title="RESOLVED">Bug 1654167 R</a><span>[Toolkit:Form Autofill]</span><span>Add preferences telemetry for Credit Card A/B testing</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653162" title="RESOLVED">Bug 1653162 R</a><span>[Toolkit:Form Autofill]</span><span>Add credit-card form telemetry for Credit Card A/B testing</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1653083" title="RESOLVED">Bug 1653083 R</a><span>[Toolkit:Form Autofill]</span><span>Add doorhanger telemetry for Credit Card A/B testing</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1655243" title="RESOLVED">Bug 1655243 R</a><span>[Toolkit:Password Manager]</span><span>Intermittent toolkit/components/passwordmgr/test/mochitest/test_DOMInputPasswordAdded_fired_between_DOMContentLoaded_and_load_events.html | [SimpleTest.finish()] No checks actually run. (You need to call ok(), is(), or similar functions at least once.  Ma</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1654843" title="RESOLVED">Bug 1654843 R</a><span>[Toolkit:Telemetry]</span><span>Add "How to update Glean SDK and Glean Parser" docs to in-tree documentation</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1635662" title="RESOLVED">Bug 1635662 R</a><span>[Toolkit:Telemetry]</span><span>Add ecosystemClientId to the Accounts Ecosystem Telemetry ping</span></div>
<div class="logline odd"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1635260" title="ASSIGNED">Bug 1635260 A</a><span>[Toolkit:Telemetry]</span><span>glean_parser should generate Rust APIs for metrics in Project FOG</span></div>
<div class="logline even"><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1624599" title="RESOLVED">Bug 1624599 R</a><span>[Toolkit:Themes]</span><span>Stop using `-moz-appearance: menuitem` for menupopup scrollbuttons on Mac</span></div>
<div class="logline odd"><span>###########</span></div>
<div class="logline even"><span>Changes to 64 bugs - 53 fixed.</span></div>
<div class="logline odd"><span>1 bugs are not included due to being restricted</span></div>
</div>
</body>
</html>
